+++ app/src/main/kotlin/com/softteco/template/data/bluetooth/BluetoothHelper.kt
@@ -1,8 +1,8 @@
-package com.softteco.template.ui.feature.bluetooth
+package com.softteco.template.data.bluetooth
 
 import android.bluetooth.BluetoothDevice
 import com.softteco.template.MainActivity
-import com.softteco.template.data.bluetooth.DataLYWSD03MMC
+import com.softteco.template.data.bluetooth.entity.DataLYWSD03MMC
 import no.nordicsemi.android.support.v18.scanner.ScanResult
 
 interface BluetoothHelper {
