+++ app/src/main/kotlin/com/softteco/template/data/bluetooth/entity/BluetoothState.kt
@@ -1,4 +1,4 @@
-package com.softteco.template.data.bluetooth
+package com.softteco.template.data.bluetooth.entity
 
 import no.nordicsemi.android.support.v18.scanner.ScanResult
