+++ app/src/main/kotlin/com/softteco/template/ui/feature/chart/ChartViewModel.kt
@@ -4,9 +4,9 @@ import android.annotation.SuppressLint
 import androidx.compose.runtime.Immutable
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.softteco.template.data.bluetooth.DataLYWSD03MMC
+import com.softteco.template.data.bluetooth.entity.DataLYWSD03MMC
 import com.softteco.template.ui.components.SnackBarState
-import com.softteco.template.ui.feature.bluetooth.BluetoothHelper
+import com.softteco.template.data.bluetooth.BluetoothHelper
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
